Better syntax

+++ artix/.local/bin/autostart.sh
@@ -1,9 +1,6 @@
 #!/usr/bin/bash
 
-  if [[ "$1" == "sway" ]]; then
-    rm -rf /home/$(whoami)/.config/sway/exec
-    touch /home/$(whoami)/.config/sway/exec
-  fi
+  if [[ "$1" == "sway" ]]; then rm -rf /home/$(whoami)/.config/sway/exec; touch /home/$(whoami)/.config/sway/exec; fi
   for service in yambar foot pipewire mako syncthing wl-paste; do
     if [[ "$1" == "sway" ]]; then
       if ! [[ $(pidof $service) ]]; then
@@ -19,29 +16,19 @@ EOF
           cat << EOF | tee -a /home/$(whoami)/.config/sway/exec > /dev/null
 exec_always syncthing serve --no-browser
 EOF
-        elif [[ "$service" == "yambar" ]]; then
-          :
+        elif [[ "$service" == "yambar" ]]; then :;
         else
           cat << EOF | tee -a /home/$(whoami)/.config/sway/exec > /dev/null
 exec_always $service
 EOF
         fi
-      elif [[ $(pidof $service) ]]; then
-        :
-      fi
+      elif [[ $(pidof $service) ]]; then :; fi
     elif [[ "$1" == "river" ]]; then
       if ! [[ $(pidof $service) ]]; then
-        if [[ "$service" == "foot" ]]; then
-          riverctl spawn "foot --server"
-        elif [[ "$service" == "wl-paste" ]]; then
-          riverctl spawn "wl-paste -t text --watch clipman store"
-        elif [[ "$service" == "syncthing" ]]; then
-          riverctl spawn "syncthing serve --no-browser"
-        else
-          riverctl spawn $service
-        fi
-      elif [[ $(pidof $service) ]]; then
-        :
-      fi
+        if [[ "$service" == "foot" ]]; then riverctl spawn "foot --server";
+        elif [[ "$service" == "wl-paste" ]]; then riverctl spawn "wl-paste -t text --watch clipman store";
+        elif [[ "$service" == "syncthing" ]]; then riverctl spawn "syncthing serve --no-browser";
+        else riverctl spawn $service; fi
+      elif [[ $(pidof $service) ]]; then :; fi
     fi
   done
